removed Abstract recorder
