+++ spynnaker/pyNN/models/common/__init__.py
@@ -1,6 +1,5 @@
 from .abstract_neuron_recordable import AbstractNeuronRecordable
 from .abstract_spike_recordable import AbstractSpikeRecordable
-from .abstract_uint32_recorder import AbstractUInt32Recorder
 from .eieio_spike_recorder import EIEIOSpikeRecorder
 from .neuron_recorder import NeuronRecorder
 from .multi_spike_recorder import MultiSpikeRecorder
@@ -9,9 +8,8 @@ from .recording_utils import get_buffer_sizes, get_data, \
 from .simple_population_settable import SimplePopulationSettable
 from .spike_recorder import SpikeRecorder
 
-__all__ = ["AbstractNeuronRecordable", "AbstractUInt32Recorder",
-           "AbstractSpikeRecordable", "EIEIOSpikeRecorder",
-           "NeuronRecorder",
+__all__ = ["AbstractNeuronRecordable", "AbstractSpikeRecordable",
+           "EIEIOSpikeRecorder", "NeuronRecorder",
            "MultiSpikeRecorder", "SimplePopulationSettable", "SpikeRecorder",
            "get_buffer_sizes", "get_data", "needs_buffering",
            "get_recording_region_size_in_bytes", "pull_off_cached_lists", ]
